[Moidfy] Voted Error Debuging
