+++ pages/dao/list/list.js
@@ -6,19 +6,19 @@ import "react-datepicker/dist/react-datepicker.css";
 import { urlFor, client } from "../../../sanity";
 import { useLoadingState } from "@assets/context/LoadingContext";
 
-import Moment from 'moment';
-
 import "react-responsive-modal/styles.css";
 import { useWalletData } from '@data/wallet';
 import { checkLogin } from "@api/UseTransactions";
 
 import toastNotify from '@utils/toast';
 import { Icon } from '@iconify/react';
-import Web3 from "web3";
 
-import { MarketContract, GovernanceContract } from "../contractHelper";
 import { voteGovernance } from "../list/useVoteGovernance"
 import { resultGovernance } from "../list/useResultGovernance"
+import { questEndTime } from "../list/useQuestEndTime"
+import { answerConfirm } from "../list/useAnswerConfirm"
+import { makeConfirm } from "../list/useMakeConfirm"
+import { cancelConfirm } from "../list/useCancelConfirm"
 
 function Index() {
   const { walletData } = useWalletData();
@@ -40,7 +40,6 @@ function Index() {
       setNowTime(new Date())
     }, 1000)
   }, [])
-  const web3 = new Web3(window.klaytn);
 
   const categories = [
     {CategoryName: 'draft'},
@@ -92,7 +91,7 @@ function Index() {
     })
   }, [newAccount]);
 
-  const GovernanceVoteHandler = async (diff, level, questKey, answer, _id, list) => {
+  const GovernanceVoteHandler = async (diff, level, questKey, answer, _id) => {
     setLoading(true);
     await voteGovernance(diff, level, questKey, answer, _id)
     setRender(!render);
@@ -100,301 +99,45 @@ function Index() {
   };
 
   const ResultHandler = async (level, _id, diff, answerKey, list) => {
-    const agreeVote = list.level === 'draft' ? list.approveTotalVote : list.level === 'success' ? list.successTotalVote : 0;
-    const disagreeVote = list.level === 'draft' ? list.rejectTotalVote : list.level === 'success' ? list.adjournTotalVote : 0;
-    const totalVote = list.level === 'draft' || list.level === 'success' ? agreeVote + disagreeVote : list.level === 'answer' ? list.answerTotalVote : 0;
-
     setLoading(true);
-
-    if(totalVote >= 10 && agreeVote === disagreeVote) {
-      try {
-        if(level === 'draft' || level === 'success') {
-          setLoading(false);
-          setSelectLevel({level: level, _id: _id})
-          setMakeSelect(true);
-        }
-      } catch (err) {
-        console.log(err);
-        setMakeSelect(false);
-        toastNotify({
-          state: 'error',
-          message: `Failed Set Quest.`,
-        });
-        setLoading(false);
-      }
-      return;
-    }
-    
-    await resultGovernance(level, _id, diff, answerKey, list)
+    await resultGovernance(level, _id, diff, answerKey, list, setSelectLevel, setMakeSelect)
     setRender(!render);
     setLoading(false);
   }
 
-  const setQuestEndTime = async (level, questKey, questId) => {
+  const setQuestEndTime = async (level, questKey, governanceId) => {
     setLoading(true);
-    const accounts = await window.klaytn.enable();
-    const account = accounts[0];
-    const endQuestQuery = `*[_type == 'governanceItem' && references('${questId}') && _id != '${Date.now()}']`
-
-    if(level === "draft") {
-      const receipt = await GovernanceContract().methods.setQuestEndTime(questKey).send({from : account})
-      console.log('draftEnd', receipt)
-      client.fetch(endQuestQuery).then(async (list) => {
-        await client.patch(list[0]._id).set({draftEndTime: Moment().format("yyyy-MM-DD HH:mm:ss")}).commit();
-        setNowTime(new Date());
-        setLoading(false);
-        setRender(!render);
-      })
-    }
-
-    if(level === "success") {
-      const receipt = await GovernanceContract().methods.setDecisionEndTime(questKey).send({from : account})
-      console.log('successEnd', receipt)
-      client.fetch(endQuestQuery).then(async (list) => {
-        await client.patch(list[0]._id).set({successEndTime: Moment().format("yyyy-MM-DD HH:mm:ss")}).commit();
-        setNowTime(new Date());
-        setLoading(false);
-        setRender(!render);
-      })
-    }
-    
-    if(level === "answer") {
-      const receipt = await GovernanceContract().methods.setAnswerEndTime(questKey).send({from : account})
-      console.log('answerEnd', receipt)
-      client.fetch(endQuestQuery).then(async (list) => {
-        await client.patch(list[0]._id).set({answerEndTime: Moment().format("yyyy-MM-DD HH:mm:ss")}).commit();
-        setNowTime(new Date());
-        setLoading(false);
-        setRender(!render);
-      })
-    }
+    await questEndTime(level, questKey, governanceId);
+    setNowTime(new Date());
+    toastNotify({
+      state: 'success',
+      message: `Success ${level} Quest Time End.`,
+    });
+    setRender(!render);
+    setLoading(false);
   }
 
   const AnswerConfirmHandler = async (questKey, answerKey, answerId, answerTitle, itemId, itemQuestId) => {
-    if(!answerKey) {
-      toastNotify({
-        state: 'error',
-        message: `Please Check Answer.`,
-      });
-      return;
-    }
+    await answerConfirm(questKey, answerKey, answerId, answerTitle, itemId, itemQuestId, setLoading, render, setRender, setSelectedAnswer);
+  }
 
-    setLoading(true);
-    try {
-      const accounts = await window.klaytn.enable();
-      const account = accounts[0];
-      const receipt = await GovernanceContract().methods.voteAnswer(questKey, answerKey).send({from : account})
-      const returnValue = receipt?.events?.VoteAnswerCast?.returnValues;
-      console.log(returnValue);
-      const answerQuery = `*[_type == 'questAnswerList' && _id == '${answerId}' && _id != '${Date.now()}']`
-      console.log('query after')
-      client.fetch(answerQuery).then(async (answer) => {
-        if(answer[0].totalVotes === null || answer[0].totalVotes === undefined) {
-          await client.patch(answer[0]._id).set({totalVotes: 0}).commit();
-          await client.patch(answer[0]._id).inc({totalVotes: returnValue.votedNfts.length}).commit();
-          await client.patch(itemId).inc({answerTotalVote: returnValue.votedNfts.length}).commit()
-        } else {
-          await client.patch(answer[0]._id).inc({totalVotes: returnValue.votedNfts.length}).commit();
-          await client.patch(itemId).inc({answerTotalVote: returnValue.votedNfts.length}).commit()
-        }
-        const SuccessAnswerQuery = `*[_type == 'governanceItemVote' && governanceItemId == '${itemQuestId}' && voter == '${account.toLowerCase()}' && _id != '${Date.now()}']`;
-        await client.fetch(SuccessAnswerQuery).then(async (list) => {
-          console.log(list);
-          await client.patch(list[0]._id).set(
-            {
-              answerOption: answerTitle,
-              answerCount: returnValue.votedNfts.length,
-              answerTxHash: receipt.transactionHash,
-            }
-          ).commit();
-        })
-        toastNotify({
-          state: 'success',
-          message: `${returnValue.votedNfts.length} Votes to Success this Quest.`,
-        });
-        setSelectedAnswer('');
-        setRender(!render);
-        setLoading(false);
-      })
-    } catch(err) {
-      console.log(err)
-      toastNotify({
-        state: 'error',
-        message: `Check Your Account.`,
-      });
-      setSelectedAnswer('');
-      setLoading(false);
+  const SelectAnswerHandler = (answer) => {
+    if(!selectedAnswer) {
+      setSelectedAnswer(answer)
+    } else {
+      setSelectedAnswer('')
     }
   }
 
   const setMakeHandler = async (e) => {
-    setMakeSelect(false);
-    setLoading(true);
-    try {
-      const newDraftTotalQuery = `*[_type == 'governanceItem' && references('${selectLevel._id}') && _id != '${Date.now()}']
-      {
-        ...,
-        'questKey': *[_type == 'quests' && _id == ^.questKey._ref && _id != '${Date.now()}'][0]
-      }`;
-      client.fetch(newDraftTotalQuery).then(async (vote) => {
-        const marketKey = vote[0].questKey.questKey;
-        const creator = vote[0].questKey.creatorAddress;
-        const title = vote[0].questKey.titleKR;
-        const creatorFee = Number(vote[0].questKey.creatorPay) / 10 ** 18;
-        const creatorFeePercentage = vote[0].questKey.creatorFee;
-        const cojamFeePercentage = vote[0].questKey.cojamFee;
-        const charityFeePercentage = vote[0].questKey.charityFee;
-        const answer = e.target.innerText.toLowerCase();
-        const questKey = vote[0].questKey.questKey;
-        const questId = vote[0]._id;
-
-        const answerKeyQuery = `*[_type == 'questAnswerList' && questKey == ${questKey} && _id != '${Date.now()}']`;
-        const answerKeyList = [];
-        await client.fetch(answerKeyQuery).then((answers) => {
-          console.log(answers);
-          answers.forEach((answer) => {
-            console.log(answer)
-            answerKeyList.push(answer.questAnswerKey);
-          });
-        });
-        console.log(selectLevel)
-        try {
-          if(selectLevel.level === 'draft') {
-            const receipt = await GovernanceContract().methods.makeQuestResult(questKey, answer).send({from : newAccount, gas: 500000})
-            console.log('makeQuest', receipt);
-            if(receipt.events.QuestResult.returnValues.result === 'approve') {
-              await client.patch(questId).set({level: 'draftEnd'}).commit();
-              const publish = await MarketContract().methods.publishMarket(
-                marketKey,
-                creator,
-                title,
-                creatorFee,
-                creatorFeePercentage,
-                cojamFeePercentage,
-                charityFeePercentage,
-                answerKeyList
-              ).send({from : newAccount, gas: 750000});
-              const approveListQuery = `*[_type == 'quests' && _id == '${vote[0].questKey._id}' && _id != '${Date.now()}']`
-              client.fetch(approveListQuery).then(async (list) => {
-                console.log(list);
-                await client.patch(list[0]._id).set({
-                  statusType: 'APPROVE',
-                  questStatus: 'APPROVE',
-                  approveTx: publish.transactionHash,
-                  approveDateTime: Moment().format("yyyy-MM-DD HH:mm:ss"),
-                  updateMember: newAccount,
-                }).commit();
-              })
-              toastNotify({
-                state: 'success',
-                message: `Approve Draft End Quest`,
-              });
-            } else {
-              await client.patch(questId).set({level: 'reject'}).commit();
-              toastNotify({
-                state: 'success',
-                message: `Reject Draft End Quest`,
-              });
-            }
-          }
-
-          if(selectLevel.level === 'success') {
-            try {
-              const receipt = await GovernanceContract().methods.makeDecisionAndExecuteAnswer(questKey, answer, answerKeyList).send({from : newAccount, gas: 500000})
-              console.log(receipt)
-              if(receipt.events.DecisionResult.returnValues.result === 'success') {
-                await client.patch(questId).set({
-                  level: 'answer',
-                  answerStartTime: Moment().format("yyyy-MM-DD HH:mm:ss"),
-                  answerEndTime: Moment().add(1, 'days').format("yyyy-MM-DD HH:mm:ss"),
-                  answerTotalVote: 0
-                }).commit();
-                toastNotify({
-                  state: 'success',
-                  message: `Success Success End Quest`,
-                });
-              } else {
-                const adjourn = await MarketContract().methods.adjournMarket(questKey).send({from : newAccount, gas: 500000})
-                console.log(adjourn)
-                await client.patch(questId).set({level: 'adjourn'}).commit();
-                const adjournListQuery = `*[_type == 'quests' && _id == '${vote[0].questKey._id}' && _id != '${Date.now()}']`
-                client.fetch(adjournListQuery).then(async (list) => {
-                  console.log(list);
-                  await client.patch(list[0]._id).set({
-                    statusType: 'ADJOURN',
-                    questStatus: 'ADJOURN',
-                    approveTx: receipt.transactionHash,
-                    adjournDateTime: Moment().format("yyyy-MM-DD HH:mm:ss"),
-                    updateMember: newAccount,
-                  }).commit();
-                })
-                toastNotify({
-                  state: 'success',
-                  message: `Success Adjourn End Quest`,
-                });
-              }
-            } catch (err) {
-              console.log(err);
-            }
-          }
-          setRender(!render);
-          setLoading(false);
-        } catch (err) {
-          console.log(err);
-          toastNotify({
-            state: 'error',
-            message: `Failed Make Quest`,
-          });
-          setLoading(false);
-        }
-      })
-    } catch (err) {
-      console.log(err)
-      toastNotify({
-        state: 'error',
-        message: `Failed Make Quest`,
-      });
-      setLoading(false);
-    }
+    await makeConfirm(e, selectLevel, setLoading, render, setRender, setMakeSelect);
   }
 
   const CancelHandler = async (list, diff) => {
-    console.log(list)
     const questKey = list.quest.questKey
-    setLoading(true);
-    if(diff < 0) {
-      try {
-        const receipt = await GovernanceContract().methods.cancelAnswer(questKey, '').send({from: newAccount, gas: 500000})
-        console.log(receipt)
-        if(receipt) {
-          const adjourn = await MarketContract().methods.adjournMarket(questKey).send({from : newAccount, gas: 500000})
-          console.log(adjourn)
-          const approveListQuery = `*[_type == 'quests' && _id == '${list.quest._id}' && _id != '${Date.now()}']`
-          client.fetch(approveListQuery).then(async (list) => {
-            console.log('list', list)
-            await client.patch(list[0]._id).set({
-              statusType: 'ADJOURN',
-              questStatus: 'ADJOURN',
-              adjournTx: receipt.transactionHash,
-              adjournDateTime: Moment().format("yyyy-MM-DD HH:mm:ss"),
-              updateMember: newAccount,
-            }).commit();
-            toastNotify({
-              state: 'success',
-              message: `Success Cancel Quest`,
-            });
-          })
-        }
-        setLoading(false);
-      } catch (err) {
-        console.log(err);
-        toastNotify({
-          state: 'error',
-          message: `Failed Cancel Quest`,
-        });
-        setLoading(false);
-      }
-    }
+    const questId = list.quest._id
+    const governanceId = list._id
+    await cancelConfirm(diff, governanceId, questKey, questId, setLoading, render, setRender);
   }
 
   return (
@@ -512,7 +255,7 @@ function Index() {
                                 const totalVote = list.answerTotalVote;
                                 const percent = !isFinite(answer.totalVotes / totalVote) ? '0' : ((answer.totalVotes / totalVote) * 100).toFixed(2);
                                 return (
-                                  <li key={index} onClick={() => setSelectedAnswer(answer)} style={{cursor:'pointer'}} className={`${selectedAnswer == answer && 'active'}`}>
+                                  <li key={index} onClick={() => SelectAnswerHandler(answer)} style={{cursor:'pointer'}} className={`${selectedAnswer == answer && 'active'}`}>
                                     <div>{answer.title}</div>
                                     <p>
                                       {answer.totalVotes ?? 0}({percent}%)
@@ -550,7 +293,10 @@ function Index() {
                             adminAddressDB !== newAccount && (
                             (list?.level === 'draft' && list?.quest?.votingList[0]?.draftTxHash) ||
                             (list?.level === 'success' && list?.quest?.votingList[0]?.successTxHash) ||
+                            (list?.level === 'success' && !(list?.quest?.votingList[0]?.draftTxHash)) ||
                             (list?.level === 'answer' && list?.quest?.votingList[0]?.answerTxHash) ||
+                            (list?.level === 'answer' && !(list?.quest?.votingList[0]?.draftTxHash)) ||
+                            (list?.level === 'answer' && !(list?.quest?.votingList[0]?.successTxHash)) ||
                             diff < 0 || totalAmount >= 5000) ? 'vote-finish' : ''}`
                         }>
                           <div>Would you like to vote for the Quest {category}?</div>
@@ -563,12 +309,12 @@ function Index() {
                                   onClick={
                                     () => {
                                       AnswerConfirmHandler(
-                                        list.quest.questKey,
-                                        selectedAnswer.questAnswerKey,
-                                        selectedAnswer._id,
-                                        selectedAnswer.title,
-                                        list._id,
-                                        list.quest._id
+                                        list?.quest?.questKey,
+                                        selectedAnswer?.questAnswerKey,
+                                        selectedAnswer?._id,
+                                        selectedAnswer?.title,
+                                        list?._id,
+                                        list?.quest._id
                                       )
                                     }
                                   }>Confirm</button>
@@ -583,10 +329,10 @@ function Index() {
                                           () => {
                                             GovernanceVoteHandler(
                                               diff,
-                                              list.level,
-                                              list.quest.questKey,
-                                              answer.title.toLowerCase(),
-                                              list.quest._id,
+                                              list?.level,
+                                              list?.quest.questKey,
+                                              answer?.title.toLowerCase(),
+                                              list?.quest._id,
                                               list
                                             )
                                           }
@@ -594,14 +340,14 @@ function Index() {
                                         disabled={
                                           (list?.level === 'draft' && list?.quest?.votingList[0]?.draftTxHash) ||
                                           (list?.level === 'success' && list?.quest?.votingList[0]?.successTxHash) ||
-                                          (list?.level === 'answer' && list?.quest?.votingList[0]?.answerTxHash) ? true : false
+                                          (list?.level === 'success' && !(list?.quest?.votingList[0]?.draftTxHash)) ||
+                                          (list?.level === 'answer' && list?.quest?.votingList[0]?.answerTxHash) ||
+                                          (list?.level === 'answer' && !(list?.quest?.votingList[0]?.draftTxHash)) ||
+                                          (list?.level === 'answer' && !(list?.quest?.votingList[0]?.successTxHash)) ||
+                                          diff < 0 || totalAmount >= 5000 ? true : false
                                         }
                                       >
-                                        {
-                                          (list?.level === 'draft' && list?.quest?.votingList[0]?.draftTxHash) ||
-                                          (list?.level === 'success' && list?.quest?.votingList[0]?.successTxHash) ||
-                                          (list?.level === 'answer' && list?.quest?.votingList[0]?.answerTxHash) ? 'VOTED' : answer.title
-                                        }
+                                        {answer.title}
                                         {
                                           adminAddressDB === newAccount ? 
                                           (
@@ -618,7 +364,7 @@ function Index() {
                           {
                             adminAddressDB === newAccount ? (
                               <button
-                                onClick={() => setQuestEndTime(list.level, list.quest.questKey, list.quest._id)}
+                                onClick={() => setQuestEndTime(list.level, list.quest.questKey, list._id)}
                                 className="adminConfirmBtn">
                                 End
                               </button>
@@ -630,9 +376,9 @@ function Index() {
                               list.level === "answer" ? (
                                 <>
                                   <button
-                                    onClick={() => ResultHandler(list.level, list.quest._id, diff, selectedAnswer.questAnswerKey)}
+                                    onClick={() => ResultHandler(list?.level, list?.quest?._id, diff, selectedAnswer?.questAnswerKey, list)}
                                     className="adminConfirmBtn">
-                                    Confirm
+                                    Result Confirm
                                   </button>
                                   <button
                                     onClick={() => CancelHandler(list, diff)}
@@ -643,10 +389,10 @@ function Index() {
                                 </>
                               ) : (
                                 <button
-                                  onClick={() => ResultHandler(list.level, list.quest._id, diff, 'not', list)}
+                                  onClick={() => ResultHandler(list?.level, list?.quest?._id, diff, 'not', list)}
                                   className="adminConfirmBtn"
                                 >
-                                  Confirm
+                                  Result Confirm
                                 </button>
                               )
                             )
@@ -654,7 +400,8 @@ function Index() {
                             (null)
                           }
                           {
-                            adminAddressDB !== newAccount && (diff < 0 || totalAmount >= 5000) ? (
+                            adminAddressDB !== newAccount && (diff < 0 || totalAmount >= 5000) ?
+                            (
                               <div className='vote-alarm'>
                                 <p>QUEST CLOSED</p>
                               </div>
@@ -662,9 +409,18 @@ function Index() {
                             adminAddressDB !== newAccount && (
                             (list?.level === 'draft' && list?.quest?.votingList[0]?.draftTxHash) ||
                             (list?.level === 'success' && list?.quest?.votingList[0]?.successTxHash) ||
-                            (list?.level === 'answer' && list?.quest?.votingList[0]?.answerTxHash)) ? (
+                            (list?.level === 'answer' && list?.quest?.votingList[0]?.answerTxHash)) ? 
+                            (
                               <div className='vote-alarm'>
                                 <p>ALREADY VOTED</p>
+                              </div>
+                            ) :
+                            (list?.level === 'success' && !(list?.quest?.votingList[0]?.draftTxHash)) ||
+                            (list?.level === 'answer' && !(list?.quest?.votingList[0]?.draftTxHash)) ||
+                            (list?.level === 'answer' && !(list?.quest?.votingList[0]?.successTxHash)) ?
+                            (
+                              <div className='vote-alarm'>
+                                <p>NO PREVIOUS VOTE</p>
                               </div>
                             ) : (null)
                           }
